refactor: extract DataGrid row focus/edit commands into a hook
Move focusTable, openRowById, handleStartRowEdit,
handleEditSelectedRow, and the scrollAndFocusRow helper out of the
already large EditableDataGrid component into a dedicated
useDataGridRowCommands hook, matching the existing
useDataGridDelete/useDataGridRowActionMenu extraction pattern in
this directory. No behavior change.
